+++ blackfynn/__init__.py
@@ -32,4 +32,4 @@ from .models import (
 )
 
 __title__ = "blackfynn"
-__version__ = "3.0.4"
+__version__ = "3.1.0"