+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Documentation at <https://developer.blackfynn.io/python/>
 
+## 3.1.0
+
+###Added
+- Function `published()` for Dataset, returns info about publication on the Discover platform
+- Function `package_count()` for Dataset, returns the total number of Packages
+- Function `team_collaborators()` for Dataset, returns the list of teams that can access the dataset
+- Function `user_collaborators()` for Dataset, returns the list of users that can access the dataset
+- Function `owner()` for Dataset, returns the owner of the dataset
+
 ## 3.0.4
 
 ### Fixed